refactor(hush-cli): extract list_verify_info from pkg_cli
Move cmd_pkg_list, cmd_pkg_verify, and cmd_pkg_info into
pkg_cli/list_verify_info.rs, all widened to pub(super) for the
dispatcher. With verify (the last non-test consumer of the trust and
store helpers) relocated, mod.rs's remaining non-dispatch imports are
now test-only or removed entirely (build_caller_auth_headers,
tempdir_for_download, several trust fns dropped; Hash removed).

Also repoint pack.rs / yank.rs / trust.rs at the helpers' canonical
homes (super::util::*, super::auth::*) instead of relying on mod.rs
re-imports that are no longer present.
